+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Removed
+- Unused wrapper prop from BlockWrapper and update related components
+
 ### Changed
 - Renamed `SInputOTP` to `SInputOtp`.
 - Renamed `SInputOTPItem` to `SInputOtpItem`.